requirements.txt
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
requests>=2.31.0
groq>=0.5.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
torch>=2.1.0
transformers>=4.40.0
trl>=0.8.0
peft>=0.10.0
accelerate>=0.27.0
datasets>=2.18.0
huggingface_hub>=0.23.0
openenv>=0.1.0
pydantic>=2.0.0
numpy>=1.26.0
